Refactor Calculator.test.tsx to flatten test structure
